+++ CHANGELOG.rst
@@ -7,6 +7,7 @@ New features
 * Add :code:`dv` option to :code:`get_observations`
 * Add context method :code:`spawn_seed`
 * Add :code:`modeling.is_simulation_model`
+* Add :code:`symbol` attribute to :code:`ColumnInfo` class 
 
 Bugfixes
 ========